+++ src/clicker/game/manual/economy.py
@@ -4,22 +4,22 @@ from ..state import GameState
 from .definitions import MANUAL_CLICK
 
 
-def get_current_income(state: GameState) -> int:
+def get_current_income(state: GameState) -> float:
     base_income = MANUAL_CLICK.base_income
     income_upgrade = state.manual_click.income_upgrade_level
     increment = MANUAL_CLICK.income_increment_per_level
 
-    return base_income + income_upgrade * increment
+    return float(base_income + income_upgrade * increment)
 
 
-def get_next_income_upgrade_cost(state: GameState) -> int:
+def get_next_income_upgrade_cost(state: GameState) -> float:
     base_cost = MANUAL_CLICK.base_income_upgrade_cost
     income_upgrade = state.manual_click.income_upgrade_level
     growth_rate = MANUAL_CLICK.growth_rate
 
     upgrade_cost = base_cost * (growth_rate ** income_upgrade)
 
-    return int(upgrade_cost)
+    return float(upgrade_cost)
 
 
 def try_manual_click(state: GameState) -> bool:
